fix: changing sku field to unique, product Post return to 201
