+++ src/main/java/com/joelmaciel/salesreport/api/controller/ProductController.java
@@ -1,13 +1,11 @@
 package com.joelmaciel.salesreport.api.controller;
 
-import com.joelmaciel.salesreport.api.controller.dto.request.ProductRequestDTO;
-import com.joelmaciel.salesreport.api.controller.dto.response.ProductDTO;
+import com.joelmaciel.salesreport.api.dto.request.ProductRequestDTO;
+import com.joelmaciel.salesreport.api.dto.response.ProductDTO;
 import com.joelmaciel.salesreport.domain.service.ProductService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
@@ -19,6 +17,7 @@ public class ProductController {
     private final ProductService productService;
 
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public ProductDTO save(@RequestBody @Valid ProductRequestDTO product) {
         return productService.save(product);
     }
