+++ src/main/java/com/joelmaciel/salesreport/api/dto/response/ProductDTO.java
@@ -1,32 +1,24 @@
-package com.joelmaciel.salesreport.api.controller.dto.request;
+package com.joelmaciel.salesreport.api.dto.response;
 
 import com.joelmaciel.salesreport.domain.model.Product;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.PositiveOrZero;
 import java.math.BigDecimal;
 
 @Builder
 @Getter
 @Setter
-public class ProductRequestDTO {
+public class ProductDTO {
 
-    @NotBlank
     private String name;
-    @NotBlank
     private String description;
-    @PositiveOrZero
-    @NotNull
     private BigDecimal price;
-    @NotBlank
     private String sku;
 
-    public static Product toModel(ProductRequestDTO product) {
-        return Product.builder()
+    public static ProductDTO toDTO(Product product) {
+        return ProductDTO.builder()
                 .name(product.getName())
                 .description(product.getDescription())
                 .price(product.getPrice())
