+++ src/main/java/com/joelmaciel/salesreport/domain/service/ProductService.java
@@ -1,7 +1,7 @@
 package com.joelmaciel.salesreport.domain.service;
 
-import com.joelmaciel.salesreport.api.controller.dto.request.ProductRequestDTO;
-import com.joelmaciel.salesreport.api.controller.dto.response.ProductDTO;
+import com.joelmaciel.salesreport.api.dto.request.ProductRequestDTO;
+import com.joelmaciel.salesreport.api.dto.response.ProductDTO;
 import com.joelmaciel.salesreport.domain.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
