+++ src/main/resources/db/migration/V002__alter-field-sku-unique.sql
@@ -1,0 +1,2 @@
+ALTER TABLE product
+ADD CONSTRAINT sku_unique_constraint UNIQUE (sku);
